requirements.txt
asgiref==3.8.1
Django==5.1.3
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
graphviz==0.20.3
inflection==0.5.1
packaging==24.2
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
